+++ shared/packages/expectationManager/src/expectationTracker/expectationTracker.ts
@@ -1,25 +1,19 @@
-// eslint-disable-next-line node/no-extraneous-import
-import { ExpectedPackageStatusAPI } from '@sofie-automation/shared-lib/dist/package-manager/package'
-import {
-	HelpfulEventEmitter,
-	LoggerInstance,
-	Reason,
-	ReturnTypeIsExpectationReadyToStartWorkingOn,
-	StatusCode,
-} from '@sofie-package-manager/api'
-import _ from 'underscore'
+import { HelpfulEventEmitter, LoggerInstance, Reason, StatusCode } from '@sofie-package-manager/api'
 import { EvaluationScheduler } from './lib/evaluationScheduler'
 import { ExpectationManagerCallbacks } from '../expectationManager'
 import { ListeningExpectationsStorage } from './lib/listeningExpectationsStorage'
 import { TrackedExpectationsStorage } from './lib/trackedExpectationsStorage'
 import { TrackedPackageContainersStorage } from './lib/trackedPackageContainersStorage'
 import { ExpectationTrackerConstants } from '../lib/constants'
-import { WorkerAgentAPI } from '../workerAgentApi'
 import { WorkInProgressTracker } from './lib/workInProgressTracker'
 import { TrackedReceivedUpdates } from './lib/trackedReceivedUpdates'
-import { InternalManager } from '../expectationManager/internalManager'
-import { expLabel, TrackedExpectation } from '../lib/trackedExpectation'
+import { InternalManager } from '../internalManager/internalManager'
+import { TrackedExpectation } from '../lib/trackedExpectation'
 import { TrackedPackageContainerExpectation } from '../lib/trackedPackageContainerExpectation'
+import { WorkerScaler } from './lib/workerScaler'
+import { TrackedExpectationAPI } from './lib/trackedExpectationAPI'
+import { TrackedPackageContainerAPI } from './lib/trackedPackageContainerAPI'
+import { TrackedPackageContainerPackageAPI } from './lib/trackedPackageContainerPackageAPI'
 
 /**
  * The ExpectationTracker is responsible for tracking and uptating the state of the Expectations
@@ -33,7 +27,8 @@ export class ExpectationTracker extends HelpfulEventEmitter {
 
 	/** This is the main store of all Tracked Expectations */
 	public trackedExpectations: TrackedExpectationsStorage
-	public waitingExpectations: TrackedExpectation[] = []
+
+	public scaler: WorkerScaler
 
 	public trackedPackageContainers: TrackedPackageContainersStorage
 
@@ -41,12 +36,16 @@ export class ExpectationTracker extends HelpfulEventEmitter {
 
 	public worksInProgress: WorkInProgressTracker
 
+	public trackedExpectationAPI: TrackedExpectationAPI
+	public trackedPackageContainerAPI: TrackedPackageContainerAPI
+	public trackedPackageContainerPackageAPI: TrackedPackageContainerPackageAPI
+
 	private logger: LoggerInstance
 	constructor(
 		private manager: InternalManager,
 		logger: LoggerInstance,
 		constants: ExpectationTrackerConstants,
-		private callbacks: ExpectationManagerCallbacks
+		public callbacks: ExpectationManagerCallbacks
 	) {
 		super()
 		this.logger = logger.category('Tracker')
@@ -54,10 +53,15 @@ export class ExpectationTracker extends HelpfulEventEmitter {
 
 		this.scheduler = new EvaluationScheduler(this.logger, this.manager, this)
 		this.worksInProgress = new WorkInProgressTracker(this.logger, this)
+		this.scaler = new WorkerScaler(this.logger, this.manager, this)
 
 		this.listeningExpectations = new ListeningExpectationsStorage(this.logger, this)
 		this.trackedExpectations = new TrackedExpectationsStorage(this)
 		this.trackedPackageContainers = new TrackedPackageContainersStorage()
+
+		this.trackedExpectationAPI = new TrackedExpectationAPI(this.logger, this)
+		this.trackedPackageContainerAPI = new TrackedPackageContainerAPI(this)
+		this.trackedPackageContainerPackageAPI = new TrackedPackageContainerPackageAPI(this)
 
 		this.receivedUpdates = new TrackedReceivedUpdates()
 	}
@@ -69,160 +73,13 @@ export class ExpectationTracker extends HelpfulEventEmitter {
 		this.trackedExpectations.clear()
 		this.trackedPackageContainers.clear()
 
-		this.scheduler.triggerEvaluateExpectations(true)
-	}
-	public triggerEvaluateExpectationsNow(): void {
-		this.scheduler.triggerEvaluateExpectations(true)
-	}
-	/** Update the state and status of a trackedExpectation */
-	public updateTrackedExpectationStatus(
-		trackedExp: TrackedExpectation,
-		upd: {
-			/** If set, sets a new state of the Expectation */
-			state?: ExpectedPackageStatusAPI.WorkStatusState
-			/** If set, sets a new reason of the Expectation */
-			reason?: Reason
-			/** If set, sets new status properties of the expectation */
-			status?: Partial<TrackedExpectation['status']> | undefined
-			/** Whether the new state is due an error or not */
-			isError?: boolean
-			/**
-			 * If set, the package on packageContainer status won't be updated.
-			 * This is used to defer updates in situations where we don't really know what the status of the package is.
-			 * */
-			dontUpdatePackage?: boolean
-		}
-	): void {
-		const { state, reason, status, isError, dontUpdatePackage } = upd
-
-		trackedExp.lastEvaluationTime = Date.now()
-		if (isError) {
-			trackedExp.lastError = {
-				time: Date.now(),
-				reason: reason || { user: 'Unknown error', tech: 'Unknown error' },
-			}
-			if (trackedExp.session) trackedExp.session.hadError = true
-		}
-
-		const prevState: ExpectedPackageStatusAPI.WorkStatusState = trackedExp.state
-		const prevReason: Reason = trackedExp.reason
-
-		let updatedState = false
-		let updatedReason = false
-		let updatedStatus = false
-
-		if (state !== undefined && trackedExp.state !== state) {
-			trackedExp.state = state
-			updatedState = true
-		}
-
-		if (reason && !_.isEqual(trackedExp.reason, reason)) {
-			trackedExp.reason = reason
-			updatedReason = true
-		}
-		if (updatedState || updatedReason) {
-			trackedExp.prevStatusReasons[prevState] = {
-				user: prevReason.user,
-				tech: `${prevReason.tech} | ${new Date().toLocaleTimeString()}`,
-			}
-		}
-
-		if (status) {
-			const newStatus = Object.assign({}, trackedExp.status, status) // extend with new values
-			if (!_.isEqual(trackedExp.status, newStatus)) {
-				Object.assign(trackedExp.status, status)
-				updatedStatus = true
-			}
-		}
-		// Log and report new states an reasons:
-		if (updatedState && updatedReason) {
-			this.logger.debug(
-				`${expLabel(trackedExp)}: New state: "${prevState}"->"${trackedExp.state}", reason: "${
-					trackedExp.reason.tech
-				}"`
-			)
-		} else if (updatedState) {
-			this.logger.debug(`${expLabel(trackedExp)}: New state: "${prevState}"->"${trackedExp.state}"`)
-		} else if (updatedReason) {
-			this.logger.debug(
-				`${expLabel(trackedExp)}: State: "${trackedExp.state}", New reason: "${trackedExp.reason.tech}"`
-			)
-		}
-
-		if (updatedState || updatedReason || updatedStatus) {
-			this.callbacks.reportExpectationStatus(trackedExp.id, trackedExp.exp, null, {
-				progress: trackedExp.status.workProgress || 0,
-				priority: trackedExp.exp.priority,
-				status: updatedState || updatedReason ? trackedExp.state : undefined,
-				statusReason: updatedReason ? trackedExp.reason : undefined,
-				prevStatusReasons: trackedExp.prevStatusReasons,
-			})
-		}
-		if (!dontUpdatePackage) {
-			if (updatedState || updatedReason || updatedStatus) {
-				this.manager.updatePackageContainerPackageStatus(trackedExp, false)
-			}
-		}
-	}
-	/** Calls workerAgent.isExpectationReadyToStartWorkingOn() */
-	public async isExpectationReadyToStartWorkingOn(
-		workerAgent: WorkerAgentAPI,
-		trackedExp: TrackedExpectation
-	): Promise<ReturnTypeIsExpectationReadyToStartWorkingOn> {
-		// First check if the Expectation depends on the fullfilled-status of another Expectation:
-		const waitingFor = this.isExpectationWaitingForOther(trackedExp)
-
-		if (waitingFor) {
-			return {
-				ready: false,
-				reason: {
-					user: `Waiting for "${waitingFor.exp.statusReport.label}"`,
-					tech: `Waiting for "${waitingFor.exp.statusReport.label}"`,
-				},
-				isWaitingForAnother: true,
-			}
-		}
-
-		return workerAgent.isExpectationReadyToStartWorkingOn(trackedExp.exp)
-	}
-	/** Checks if the expectation is waiting for another expectation, and returns the awaited Expectation, otherwise null */
-	public isExpectationWaitingForOther(trackedExp: TrackedExpectation): TrackedExpectation | null {
-		if (trackedExp.exp.dependsOnFullfilled?.length) {
-			// Check if those are fullfilled:
-			let waitingFor: TrackedExpectation | undefined = undefined
-			for (const id of trackedExp.exp.dependsOnFullfilled) {
-				const trackedExp = this.trackedExpectations.get(id)
-				if (trackedExp && trackedExp.state !== ExpectedPackageStatusAPI.WorkStatusState.FULFILLED) {
-					waitingFor = trackedExp
-					break
-				}
-			}
-			if (waitingFor) {
-				return waitingFor
-			}
-		}
-		return null
+		this.scheduler.triggerEvaluation(true)
 	}
 	/**
-	 * To be called when trackedExp.status turns fullfilled.
-	 * Triggers any other expectations that listens to (are dependant on) the fullfilled one.
-	 * @returns true if any other expectations where triggered (ie evaluation should run again ASAP)
+	 * Set the scheduler to trigger another evaluation ASAP
 	 */
-	public onExpectationFullfilled(fullfilledExp: TrackedExpectation): boolean {
-		let hasTriggeredSomething = false
-		if (fullfilledExp.state === ExpectedPackageStatusAPI.WorkStatusState.FULFILLED) {
-			const expectationsToTrigger = this.listeningExpectations.getListeningExpectations(fullfilledExp.id)
-
-			// Go through the listening expectations and mark them to be re-evaluated ASAP:
-			for (const id of expectationsToTrigger) {
-				const toTriggerExp = this.trackedExpectations.get(id)
-				if (toTriggerExp) {
-					toTriggerExp.lastEvaluationTime = 0 // so that it reruns ASAP
-					hasTriggeredSomething = true
-				}
-			}
-		}
-		return hasTriggeredSomething
+	public triggerEvaluationNow(): void {
+		this.scheduler.triggerEvaluation(true)
 	}
 
 	/** Returns the appropriate time to wait before checking a fulfilled expectation again */
@@ -248,7 +105,7 @@ export class ExpectationTracker extends HelpfulEventEmitter {
 		}
 		trackedPackageContainer.status.statusChanged = Date.now()
 
-		this.manager.updateTrackedPackageContainerMonitorStatus(
+		this.trackedPackageContainerAPI.updateTrackedPackageContainerMonitorStatus(
 			trackedPackageContainer,
 			monitorId,
 			undefined,
@@ -259,137 +116,7 @@ export class ExpectationTracker extends HelpfulEventEmitter {
 	public getTrackedPackageContainers(): TrackedPackageContainerExpectation[] {
 		return Object.values(this.trackedPackageContainers)
 	}
-
-	/**
-	 * Goes through the expectations and checks if there are too many expectations waiting for workers
-	 * and tries to scale up new workers if needed.
-	 * Called when a pass of evaluating expectation has finished
-	 */
-	public async checkIfNeedToScaleUp(): Promise<void> {
-		const waitingExpectations: TrackedExpectation[] = []
-		const waitingPackageContainers: TrackedPackageContainerExpectation[] = []
-
-		let requestsSentCount = 0
-
-		for (const exp of this.trackedExpectations.list()) {
-			/** The expectation is waiting for a worker */
-			const isWaiting: boolean =
-				exp.state === ExpectedPackageStatusAPI.WorkStatusState.NEW ||
-				exp.state === ExpectedPackageStatusAPI.WorkStatusState.WAITING ||
-				exp.state === ExpectedPackageStatusAPI.WorkStatusState.READY
-
-			/** Not supported by any worker */
-			const notSupportedByAnyWorker: boolean = Object.keys(exp.availableWorkers).length === 0
-			/** No worker has had time to work on it lately */
-			const notAssignedToAnyWorker: boolean =
-				!!exp.noWorkerAssignedTime && Date.now() - exp.noWorkerAssignedTime > this.constants.SCALE_UP_TIME
-
-			if (
-				isWaiting &&
-				(notSupportedByAnyWorker || notAssignedToAnyWorker) &&
-				!this.isExpectationWaitingForOther(exp) // Filter out expectations that aren't ready to begin working on anyway
-			) {
-				if (!exp.waitingForWorkerTime) {
-					this.logger.silly(
-						`Starting to track how long expectation "${expLabel(exp)}" has been waiting for a worker`
-					)
-					exp.waitingForWorkerTime = Date.now()
-				}
-			} else {
-				exp.waitingForWorkerTime = null
-			}
-			if (exp.waitingForWorkerTime) {
-				const hasBeenWaitingFor = Date.now() - exp.waitingForWorkerTime
-				if (
-					hasBeenWaitingFor > this.constants.SCALE_UP_TIME || // Don't scale up too fast
-					this.manager.workerAgents.list().length === 0 // Although if there are no workers connected, we should scale up right away
-				) {
-					waitingExpectations.push(exp)
-				} else {
-					this.logger.silly(
-						`Expectation "${expLabel(exp)}" has been waiting for less than ${
-							this.constants.SCALE_UP_TIME
-						}ms (${hasBeenWaitingFor}ms), letting it wait a bit more`
-					)
-				}
-			}
-		}
-		for (const exp of waitingExpectations) {
-			if (requestsSentCount < this.constants.SCALE_UP_COUNT) {
-				requestsSentCount++
-				this.logger.debug(`Requesting more resources to handle expectation "${expLabel(exp)}"`)
-				await this.manager.workforceAPI.requestResourcesForExpectation(exp.exp)
-			}
-		}
-
-		this.logger.silly(`Waiting expectations: ${waitingExpectations.length}, sent out ${requestsSentCount} requests`)
-
-		for (const packageContainer of this.getTrackedPackageContainers()) {
-			if (!packageContainer.currentWorker) {
-				if (!packageContainer.waitingForWorkerTime) {
-					packageContainer.waitingForWorkerTime = Date.now()
-				}
-			} else {
-				packageContainer.waitingForWorkerTime = null
-			}
-			if (
-				packageContainer.waitingForWorkerTime &&
-				Date.now() - packageContainer.waitingForWorkerTime > this.constants.SCALE_UP_TIME
-			) {
-				waitingPackageContainers.push(packageContainer)
-			}
-		}
-		for (const packageContainer of waitingPackageContainers) {
-			if (requestsSentCount < this.constants.SCALE_UP_COUNT) {
-				requestsSentCount++
-				this.logger.debug(`Requesting more resources to handle packageContainer "${packageContainer.id}"`)
-				await this.manager.workforceAPI.requestResourcesForPackageContainer(packageContainer.packageContainer)
-			}
-		}
-
-		this.waitingExpectations = waitingExpectations
-	}
 	public getSortedTrackedExpectations(): TrackedExpectation[] {
 		return this.trackedExpectations.list()
-	}
-	/**
-	 * Handle an Expectation that had no worker assigned
-	 */
-	public noWorkerAssigned(trackedExp: TrackedExpectation): void {
-		if (!trackedExp.session) throw new Error('Internal error: noWorkerAssigned: session not set')
-		if (trackedExp.session.assignedWorker)
-			throw new Error('Internal error: noWorkerAssigned can only be called when assignedWorker is falsy')
-
-		let noAssignedWorkerReason: ExpectedPackageStatusAPI.Reason
-		if (!trackedExp.session.noAssignedWorkerReason) {
-			this.logger.error(
-				`trackedExp.session.noAssignedWorkerReason is undefined, although assignedWorker was set..`
-			)
-			noAssignedWorkerReason = {
-				user: 'Unknown reason (internal error)',
-				tech: 'Unknown reason',
-			}
-		} else {
-			noAssignedWorkerReason = trackedExp.session.noAssignedWorkerReason
-		}
-
-		if (!trackedExp.noWorkerAssignedTime) trackedExp.noWorkerAssignedTime = Date.now()
-
-		// Special case: When WAITING and no worker was assigned, return to NEW so that another worker might be assigned:
-		if (trackedExp.state === ExpectedPackageStatusAPI.WorkStatusState.WAITING) {
-			this.updateTrackedExpectationStatus(trackedExp, {
-				state: ExpectedPackageStatusAPI.WorkStatusState.NEW,
-				reason: noAssignedWorkerReason,
-				// Don't update the package status, since this means that we don't know anything about the package:
-				dontUpdatePackage: true,
-			})
-		} else {
-			// only update the reason
-			this.updateTrackedExpectationStatus(trackedExp, {
-				reason: noAssignedWorkerReason,
-				// Don't update the package status, since this means that we don't know anything about the package:
-				dontUpdatePackage: true,
-			})
-		}
 	}
 }
